backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-login
pandas
numpy
scikit-learn
joblib
python-dotenv
pymysql
cryptography
gunicorn
bcrypt
